# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FedRAMP Gap Analysis Agent API
  version: 1.0.0
  description: |
    AI-powered FedRAMP High baseline compliance gap analysis agent that analyzes policy documents, 
    design documentation, and source code to identify compliance gaps and provide remediation recommendations.

    ## Features
    - Comprehensive gap analysis across policy, design, and code
    - Support for 325+ FedRAMP High baseline controls (NIST 800-53 Rev 5)
    - Risk assessment with severity classification
    - Automated remediation recommendations with code examples
    - Multiple report formats (JSON, PDF, HTML, Excel)

    ## Authentication
    All API endpoints require authentication using Bearer token or API key.

  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.fedramp-agent.example.com/api/v1
    description: Production server
  - url: https://staging-api.fedramp-agent.example.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
  /analyze/comprehensive:
    post:
      tags:
        - Analysis
      summary: Start comprehensive gap analysis
      description: |
        Initiates a comprehensive FedRAMP gap analysis that includes:
        1. Policy document parsing
        2. Design document analysis
        3. Code repository scanning
        4. Gap detection
        5. Risk assessment
        6. Remediation recommendations
      operationId: startComprehensiveAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComprehensiveAnalysisRequest"
      responses:
        "202":
          description: Analysis job started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /analyze/policy:
    post:
      tags:
        - Analysis
      summary: Analyze policy documents
      description: Extract and analyze FedRAMP requirements from policy documents
      operationId: analyzePolicyDocuments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyAnalysisRequest"
      responses:
        "202":
          description: Policy analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      security:
        - BearerAuth: []
  /analyze/code:
    post:
      tags:
        - Analysis
      summary: Analyze code repository
      description: Scan source code for security patterns and compliance implementations
      operationId: analyzeCodeRepository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeAnalysisRequest"
      responses:
        "202":
          description: Code analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      security:
        - BearerAuth: []
  /jobs/{jobId}:
    get:
      tags:
        - Analysis
      summary: Get job status
      description: Retrieve the current status and progress of an analysis job
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: Job identifier
      responses:
        "200":
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []

    delete:
      tags:
        - Analysis
      summary: Cancel analysis job
      description: Cancel a running analysis job
      operationId: cancelJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Job cancelled successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
  /jobs/{jobId}/result:
    get:
      tags:
        - Analysis
      summary: Get analysis results
      description: Retrieve complete analysis results for a completed job
      operationId: getAnalysisResult
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis results retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResultResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
    ComprehensiveAnalysisRequest:
      type: object
      required:
        - policy_documents
      properties:
        policy_documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentInput"
          description: Policy documents containing FedRAMP requirements
        design_documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentInput"
          description: Design documents with system specifications
        repository:
          $ref: "#/components/schemas/RepositoryInput"
        analysis_options:
          $ref: "#/components/schemas/AnalysisOptions"
        webhook_url:
          type: string
          format: uri
          description: Webhook URL for completion notification
    DocumentInput:
      type: object
      required:
        - url
        - type
      properties:
        url:
          type: string
          description: URL or path to the document
        type:
          type: string
          enum: [pdf, docx, confluence, html]
          description: Document type
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata
    RepositoryInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: Git repository URL
        branch:
          type: string
          default: main
          description: Branch to analyze
        local_path:
          type: string
          description: Local repository path (if already cloned)
        credentials:
          $ref: "#/components/schemas/RepositoryCredentials"
    RepositoryCredentials:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [token, ssh, basic]
          description: Credential type
        token:
          type: string
          description: Access token
        username:
          type: string
          description: Username for basic auth
        password:
          type: string
          format: password
          description: Password for basic auth
        ssh_key:
          type: string
          description: SSH private key
    AnalysisOptions:
      type: object
      properties:
        control_families:
          type: array
          items:
            type: string
          description: Specific control families to analyze
        include_remediation:
          type: boolean
          default: true
          description: Include remediation recommendations
        include_code_examples:
          type: boolean
          default: true
          description: Include code examples in remediation
        risk_threshold:
          type: number
          minimum: 0
          maximum: 10
          description: Only report gaps with risk score above threshold
        severity_filter:
          type: array
          items:
            type: string
            enum: [critical, high, medium, low]
          description: Filter gaps by severity
    PolicyAnalysisRequest:
      type: object
      required:
        - policy_documents
      properties:
        policy_documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentInput"
        analysis_options:
          $ref: "#/components/schemas/AnalysisOptions"
    CodeAnalysisRequest:
      type: object
      required:
        - repository
      properties:
        repository:
          $ref: "#/components/schemas/RepositoryInput"
        analysis_options:
          $ref: "#/components/schemas/AnalysisOptions"
    AnalysisJobResponse:
      type: object
      required:
        - job_id
        - status
        - message
        - created_at
      properties:
        job_id:
          type: string
          description: Unique job identifier
        status:
          type: string
          enum:
            [
              pending,
              parsing_documents,
              analyzing_code,
              detecting_gaps,
              assessing_risk,
              generating_remediation,
              generating_report,
              completed,
              failed,
            ]
          description: Job status
        message:
          type: string
          description: Status message
        created_at:
          type: string
          format: date-time
          description: Job creation timestamp
        estimated_completion_time:
          type: string
          format: date-time
          description: Estimated completion time
    JobStatusResponse:
      type: object
      required:
        - job_id
        - status
        - progress
        - created_at
        - updated_at
        - result_available
      properties:
        job_id:
          type: string
        status:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 100
          description: Progress percentage
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        error_message:
          type: string
          nullable: true
        result_available:
          type: boolean
    AnalysisResultResponse:
      type: object
      required:
        - job_id
        - analysis_date
        - gaps
        - risk_assessments
        - remediations
        - summary
        - control_coverage
      properties:
        job_id:
          type: string
        analysis_date:
          type: string
          format: date-time
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/Gap"
        risk_assessments:
          type: array
          items:
            $ref: "#/components/schemas/RiskAssessment"
        remediations:
          type: array
          items:
            $ref: "#/components/schemas/Remediation"
        summary:
          $ref: "#/components/schemas/AnalysisSummary"
        control_coverage:
          $ref: "#/components/schemas/ControlCoverage"
    Gap:
      type: object
      properties:
        gap_id:
          type: string
        control_id:
          type: string
        control_name:
          type: string
        gap_type:
          type: string
          enum:
            [
              missing_in_design,
              missing_in_code,
              missing_implementation,
              partial_implementation,
              incorrect_implementation,
              outdated_implementation,
              configuration_gap,
            ]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        description:
          type: string
        policy_requirement:
          type: string
        design_specification:
          type: string
          nullable: true
        code_implementation:
          type: string
          nullable: true
        evidence:
          type: array
          items:
            type: object
        risk_score:
          type: number
          minimum: 0
          maximum: 10
        impact:
          type: string
        likelihood:
          type: string
    RiskAssessment:
      type: object
      properties:
        gap_id:
          type: string
        control_id:
          type: string
        risk_score:
          type: number
        impact:
          type: string
        likelihood:
          type: string
        risk_level:
          type: string
        business_impact:
          type: string
        technical_impact:
          type: string
        compliance_impact:
          type: string
        exploitability:
          type: number
        remediation_priority:
          type: integer
    Remediation:
      type: object
      properties:
        gap_id:
          type: string
        control_id:
          type: string
        summary:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RemediationStep"
        effort_estimate:
          type: string
        estimated_hours:
          type: integer
        required_skills:
          type: array
          items:
            type: string
        testing_guidance:
          type: string
        validation_criteria:
          type: array
          items:
            type: string
    RemediationStep:
      type: object
      properties:
        step_number:
          type: integer
        description:
          type: string
        technical_details:
          type: string
        code_example:
          type: string
          nullable: true
        verification:
          type: string
    AnalysisSummary:
      type: object
      properties:
        total_controls_evaluated:
          type: integer
        controls_with_gaps:
          type: integer
        controls_compliant:
          type: integer
        total_gaps:
          type: integer
        critical_gaps:
          type: integer
        high_gaps:
          type: integer
        medium_gaps:
          type: integer
        low_gaps:
          type: integer
        average_risk_score:
          type: number
        compliance_score:
          type: number
    ControlCoverage:
      type: object
      properties:
        total_required:
          type: integer
        total_implemented:
          type: integer
        coverage_percentage:
          type: number
        by_family:
          type: object
          additionalProperties:
            type: object
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

# Made with Bob